+++ src/app/[lang]/(home)/home/_components/SimpleReportCard.tsx
