+++ src/app/[lang]/(home)/home/news.tsx
