+++ src/middleware.ts
@@ -53,7 +53,7 @@ export const config = {
      * - _next/image (image optimization files)
      * - favicon.ico (favicon file)
      */
-    "/((?!api|_next/static|_next/image|favicon.ico|icons/.*).*)",
+    "/((?!api|_next/static|_next/image|favicon.ico|icons/.*|images/.*).*)",
   ],
 };
